[FF-A] Notification Service Updates (#339)
Updated the ID and Bitmap values to UINT32s.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?
- [ ] Backport to release branch?

Tested with the FfaPartitionTestApp

N/A

(cherry picked from commit fc8f6fd677f6de60e597b0edadeb9fab2febe544)

+++ FfaFeaturePkg/Include/Library/NotificationServiceLib.h
@@ -61,7 +61,7 @@ NotificationServiceHandle (
 **/
 NotificationStatus
 NotificationServiceIdSet (
-  UINT16  Id,
+  UINT32  Id,
   UINT8   *ServiceUuid,
   UINT32  Flag
   );
